Stop calling setMarkedText upon (de)activation
Fixes #346

Certain apps don't handle -setMarkedText:selectedRange:replacementRange:
consistently upon IMK session (de)activation. It's an effectively no-op
anyway, and so this simplifies the (de)activation call.

Also updates the state machine transitions so that entering the
deactivated state will cause the state machine to transit to the empty
state. This is to guarantee that the state machine is always in the
empty state upon IMK session activation.