Reset selectCandidateWithNumericKeypad for testing
This issue was discovered after #381 if the tests were run locally when
`selectCandidateWithNumericKeypad` is set to true. This commit makes the
test reset and restore the setting during testing.

Ideally we shouldn't let Preferences affect tests, since they are
supposed to be hermetic. We have discussed the issue in #294. Until the
code is properly refactored, though, this will make the tests pass
locally.

Also updated the reset-restore code elsewhere in the tests so that the
restore always takes place before the assertions.