Show custom size in Candidate Text Size setting
Before this treatment, if a user uses a custom candidate text size, the
pop up button will revert to the first item (the smallest font size) in
the list since we now remove the empty item at the bottom. If the user
doesn't touch the pop up button, the custom font size remains intact.
It's still a less ideal UI, though.

This commit makes sure that the custom font size is inserted into the
pop up button and is selected.