Uses the legacy UI on macOS before Big Sur
The NSVisualEffectView usage runs into issues when on macOS versions
between 10.13 and 10.15, inclusive, with behavior being different
between 10.13 on the one side and 10.14 and 10.15 on the other side.

Given the effort that might be required, we think it best to switch back
to the legacy design when run on macOS versions earlier than Big Sur
(macOS 11.0).