Sync linted C++ source with fcitx5-mcbopomofo
This carries over several fixes there.

Note: after this, the header guards will be like "SRC_ENGINE_FOO_H" in
accordance with Google C++ style. We understand that in the macOS
version, the source is located in Source/. However the source in
fcitx5-mcbopomofo is in src/. Since we do most of the linting and
tidying there, we think it makes sense to use the header guards this
way.