Reset the key label index after scrolling
If the vertical candidate table view gets scrolled, the key strip label
highlight also needs to be readjusted, otherwise the highlighted key
label won't match that of the new highlighted row in the table view.

Notice that we only readjust keyLabelStripView.highlightedIndex after
the scroll is done. We hide the highlight during the scroll, so as to
reduce the visible mismatch between the two highlights.