feat: Add option to use HL keys to move cursor in candidate window
This commit introduces a new preference that allows users to use H and L keys to move the cursor left and right in the candidate window, as an alternative to the existing J and K keys.
The preference is implemented as a dropdown menu in the preferences panel, with the options:
- Disabled
- Use JK keys
- Use HL keys
The KeyHandler is updated to respect this new setting.