Fix vertical candidate window flickering
The flickering issue is more often found with the vertical candidate
window, likely due to the layout passes it needs to go through. This PR
makes sure that candidate window position is first set before the window
becomes visible.

This PR also switches to use dispatch queue for all delayed method
invocation. This makes sure that all such deferred actions are put in
the same queue, and so they will be performed in the expected order.